HDDS-1228. Fix checkstyle/findbugs/rat, move comparator to impl package
